Another 64bit fix.
git-svn-id: file:///home/nick/dev/d/proj/svn-to-git-scrap/svn-dump/filtered-repos/SemiTwistDTools/trunk@241 183b819f-4d41-4c81-9374-59e59addc796

+++ src/semitwist/util/text.d
@@ -360,9 +360,9 @@ T stripNonPrintable(T)(T str) if(isSomeString!T)
 }
 
 /// Return value is number of code units
-uint nextCodePointSize(T)(T str) if(is(T==string) || is(T==wstring))
+size_t nextCodePointSize(T)(T str) if(is(T==string) || is(T==wstring))
 {
-	uint ret;
+	size_t ret;
 	str.decode(ret);
 	return ret;
 }
